fix(mentor-hq): pin three to r128 so the HQ world matches the prototype by construction
Stop emulating r128's render pipeline inside r185 (the whack-a-mole that kept
the world warmer/more saturated/darker than the prototype — most visibly dusk,
where the land dimmed into shadow). three is used ONLY in the HQ 3D scene
(fiber/drei already removed), so pinning it is isolated and low-risk.

- three 0.185.0 -> 0.128.0 (the exact version the prototype loads), @types/three
  0.185.0 -> 0.128.0. Reinstalled.

API calls reverted to r128 + emulation hacks removed (hqScene.ts):
- renderer.outputColorSpace = SRGBColorSpace  ->  renderer.outputEncoding = sRGBEncoding
- deleted THREE.ColorManagement.enabled = false (ColorManagement does not exist in
  r128 — the line would throw; r128 uses sRGB albedos as-authored)
- sky CanvasTexture: tx.colorSpace = SRGBColorSpace  ->  tx.encoding = sRGBEncoding
- deleted the hemi/amb intensity *= Math.PI legacy-lighting emulation (r128 applies
  that PI boost internally; keeping it would double-apply)
- kept PCFSoftShadowMap (a REAL soft shadow on r128; r185 silently downgraded it to
  PCFShadowMap = the harder/darker dusk shadows) and ACESFilmicToneMapping
- hqKit.ts: cast position.array to Float32Array for .slice() (r128 @types declare
  BufferAttribute.array as ArrayLike<number>)

TIME.golden + THEMES.mix unchanged (prototype's exact authored values) — on r128
they reproduce the look by construction, no tuning. Kept: real-data wiring, the 3
verification world-states, the SSR client-only lazy boundary (three stays out of
the Worker bundle), the dusk toggle, Beacon, index-in-place.

Verified in-browser side-by-side vs the live prototype in the SAME tab (identical
1710x951 viewport + identical home camera) in BOTH golden and dusk: now match —
same building/grass/sky tone, same brightness, same soft shadows, dusk land stays
bright under a dark sky. Gates green (tsc, lint, build).

+++ src/components/mentor-hq/world/hqKit.ts
@@ -595,6 +595,8 @@ export function cloth(w: number, h: number, color: string, o: MOpts = {}): THREE
   );
   m.castShadow = true;
   m.userData.cloth = true;
-  m.userData.base = g.attributes.position.array.slice();
+  // @types/three@0.128 types BufferAttribute.array as ArrayLike<number>; the
+  // runtime value is a Float32Array. Cast so .slice() type-checks on r128 types.
+  m.userData.base = (g.attributes.position.array as Float32Array).slice();
   return m;
 }
